feat: merge setup code into main script
