+++ src/ecr_check_repos_for_terraform/check.py
@@ -1,11 +1,31 @@
+import argparse
 import logging
 
-# initialize the app
-from ecr_check_repos_for_terraform.config import config
+import ecr_check_repos_for_terraform.context
 
 from ecr_check_repos_for_terraform.check_repos import check
+from lib.logging import setup_logging
+
+
+def parse_args():
+  ''' parse command line args '''
+  parser = argparse.ArgumentParser(
+    formatter_class=argparse.ArgumentDefaultsHelpFormatter
+  )
+  parser.add_argument(
+    '--loglevel',
+    choices=['DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL'],
+    default='INFO',
+    help='log level'
+  )
+  return parser.parse_args()
+
 
 if __name__ == "__main__":
+
+  args = parse_args()
+  loglevel = args.loglevel
+  setup_logging(eval('logging.' + loglevel))
 
   logging.info(
     'Checking ECR for repositories that perhaps should be added to Terraform...'
